+++ Assets/ProceduralStages/Terrain/Props/PropsDefinition.cs
@@ -13,6 +13,8 @@ namespace ProceduralStages
     {
         public string asset;
         public float scale = 1;
+        public float minScale = 1;
+        public float maxScale = 1;
         public bool ground;
         public int count;
         public bool changeColor;